add in sentry and update some error handling
