+++ config/env.js
@@ -1,5 +1,6 @@
 const fs = require('fs')
 const path = require('path')
+const pkg = require('../package.json')
 const paths = require('./paths')
 
 // Make sure that including paths.js after env.js will read .env variables.
@@ -73,6 +74,8 @@ function getClientEnvironment(publicUrl) {
         // This should only be used as an escape hatch. Normally you would put
         // images into the `src` and `import` them in code to get their paths.
         PUBLIC_URL: publicUrl,
+        SENTRY_DSN: process.env.SENTRY_DSN,
+        VESION: pkg.version,
       }
     )
   // Stringify all values so we can feed into Webpack DefinePlugin
