+++ public/index.html
@@ -10,6 +10,12 @@
       href="https://fonts.googleapis.com/css?family=IBM+Plex+Sans&display=swap"
       rel="stylesheet"
     />
+    <style type="text/css">
+      body,
+      html {
+        margin: 0;
+      }
+    </style>
     <script src="./buy.js"></script>
     <!-- <link rel="manifest" href="/manifest.json"> -->
     <!--
