+++ src/background/alarm/prediction.js
@@ -4,10 +4,10 @@ import {
   P_ALARM,
   MAX_ACTIVITY_ROWS,
 } from '../../constants'
+import { tracker } from '../../lib/error-tracker'
 import { last } from '../../lib/functional'
 import { api } from '../../lib/mujo-sdk'
 import { storage, getActivity, resetActivity } from '../storage'
-import { exception } from '../tracking'
 import { upsertAlarm } from './util'
 
 export const alarmKey = date => `${P_ALARM}_${date}`
@@ -33,7 +33,7 @@ export const checkPredictions = async (options = {}) => {
       await api.postActivity(activity.slice(MAX_ACTIVITY_ROWS * -1))
       await resetActivity()
     } catch (e) {
-      exception(e)
+      tracker.exception(e)
     }
   }
 
@@ -42,7 +42,7 @@ export const checkPredictions = async (options = {}) => {
     predictions = await api.getBreaks()
     await storage.set(PREDICTED_BREAK_TIMES, predictions)
   } catch (e) {
-    exception(e)
+    tracker.exception(e)
     predictions = []
   }
   const upcomingPredictions = predictions.filter(currentAlarms)
