+++ src/env.js
@@ -1,0 +1,2 @@
+export const ENVIRONMENT = process.env.NODE_ENV
+export const { SENTRY_DSN, VERSION } = process.env
