+++ src/lib/tracker.js
@@ -12,11 +12,3 @@ export const track = (options = {}, overrides = {}) => {
   const payload = { category, action, label, value, event }
   return message(TRACK, { payload })
 }
-
-export const exception = (err, overrides = {}) => {
-  const { event = 'exception' } = overrides
-  const errorStack = err.stack
-  const errorMessage = err.message
-  const payload = { errorStack, errorMessage, event }
-  return message(TRACK, { payload })
-}
